requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
flickr_api
pymongo
ipdb
python-flickr
# ecdsa
# paramiko
# pycrypto
